@@ -1,7 +1,8 @@
 import { useState, useEffect, createRef } from 'react'
-import { useHistory } from 'react-router-dom'
+import { useNavigate } from 'react-router-dom'
 import { Button } from 'primereact/button'
-import { Column, DataTable } from 'primereact/datatable'
+import { DataTable } from 'primereact/datatable'
+import { Column } from 'primereact/column'
 import { peliculaService } from '../services/peliculaService'
 import { InputText } from 'primereact/inputtext'
 import { Toast } from 'primereact/toast'
@@ -9,11 +10,11 @@ import { Toast } from 'primereact/toast'
 export function BuscarPeliculas() {
   const [peliculas, setPeliculas] = useState([])
   const [filtroBusqueda, setFiltroBusqueda] = useState('')
-  const history = useHistory()
+  const navigate = useNavigate()
   const toast = createRef()
   
   function editar(pelicula) {
-    return (<Button tooltip="Editar la película" icon="pi pi-chevron-right" className="p-button-raised p-button-info p-button-rounded" onClick={() => history.push(`editarPelicula/${pelicula.id}`)} />)
+    return <Button tooltip="Editar la película" icon="pi pi-chevron-right" className="p-button-raised p-button-info p-button-rounded" onClick={() => navigate(`editarPelicula/${pelicula.id}`)} />
   }
 
   async function eliminarPelicula(pelicula) {
@@ -29,7 +30,7 @@ export function BuscarPeliculas() {
 }
 
   function eliminar(pelicula) {
-    return (<Button tooltip="Eliminar la película" icon="pi pi-times" className="p-button-raised p-button-danger p-button-rounded" onClick={() => eliminarPelicula(pelicula)} />)
+    return <Button tooltip="Eliminar la película" icon="pi pi-times" className="p-button-raised p-button-danger p-button-rounded" onClick={() => eliminarPelicula(pelicula)} />
   }
 
   useEffect(() => {
@@ -58,10 +59,10 @@ export function BuscarPeliculas() {
         <Column field="titulo" style={{width: '25%'}} header="Título"></Column>
         <Column field="frase" header="Frase"></Column>
         <Column field="anio" style={{width: defaultButtonWidth}} header="Año"></Column>
-        <Column body={editar} style={{width: defaultButtonWidth}} />
-        <Column body={eliminar} style={{width: defaultButtonWidth}} />
+        <Column body={editar} style={{width: defaultButtonWidth}} ></Column>
+        <Column body={eliminar} style={{width: defaultButtonWidth}}></Column>
       </DataTable>
-      <Button icon="pi pi-plus" label="Agregar una nueva película" className="p-button-primary p-button-outlined p-button-rounded" onClick={() => { history.push('/crearPelicula')}}></Button>
+      <Button icon="pi pi-plus" label="Agregar una nueva película" className="p-button-primary p-button-outlined p-button-rounded" onClick={() => { navigate('/crearPelicula')}}></Button>
     </div>
   )
 }
