@@ -1,5 +1,8 @@
-import { PeliculasRoutes } from './routes'
 import './App.css'
+import 'primereact/resources/themes/saga-blue/theme.css'
+import 'primereact/resources/primereact.min.css'
+import 'primeicons/primeicons.css'
+import { PeliculasRoutes } from './routes'
 
 function App() {
   return (
