@@ -1,7 +1,7 @@
 import { render, screen } from '@testing-library/react'
 import App from './App'
 
-test('renders learn react link', () => {
+test('renders app', () => {
   render(<App />)
   const linkElement = screen.getByText(/Películas/i)
   expect(linkElement).toBeInTheDocument()
